evaluation: move tag handling urls to AuroraProject root urls.
