+++ Evaluation/views.py
@@ -654,22 +654,6 @@ def autocomplete_user(request, course_short_title=None):
 
 @aurora_login_required()
 @staff_member_required
-def autocomplete_tag(request, course_short_title=None):
-    term = request.GET.get('term', '')
-    content_type_id = request.GET['content_type_id']
-
-    content_type = ContentType.objects.get_for_id(content_type_id)
-    taggable_model = content_type.model_class()
-    tags = taggable_model.tags.all().filter(
-        Q(name__istartswith=term)
-    )
-    names = [tag.name for tag in tags]
-    response_data = json.dumps(names, ensure_ascii=False)
-    return HttpResponse(response_data, content_type='application/json; charset=utf-8')
-
-
-@aurora_login_required()
-@staff_member_required
 def load_reviews(request, course_short_title=None):
     course = Course.get_or_raise_404(short_title=course_short_title)
     if not 'elaboration_id' in request.GET:
@@ -680,6 +664,7 @@ def load_reviews(request, course_short_title=None):
 
     return render_to_response('task.html', {'elaboration': elaboration, 'reviews': reviews, 'stack': 'stack', 'course':course},
                               RequestContext(request))
+
 
 @aurora_login_required()
 @staff_member_required
@@ -695,6 +680,7 @@ def load_task(request, course_short_title=None):
 
     return render_to_response('task_s.html', {'stack_elaborations':stack_elaborations, 'elaboration': elaboration, 'reviews': reviews, 'stack': 'stack', 'course':course},
                               RequestContext(request))
+
 
 @require_POST
 @csrf_exempt
@@ -907,33 +893,6 @@ def get_points(request, user, course):
         data['stacks'].append(stack_data)
 
     return data
-
-@csrf_exempt
-@staff_member_required
-def add_tags(request, course_short_title=None):
-    text = request.POST['text']
-    object_id = request.POST['object_id']
-    content_type_id = request.POST['content_type_id']
-
-    content_type = ContentType.objects.get_for_id(content_type_id)
-    taggable_object = content_type.get_object_for_this_type(pk=object_id)
-    taggable_object.add_tags_from_text(text)
-
-    return render_to_response('tags.html', {'tagged_object': taggable_object}, context_instance=RequestContext(request))
-
-@csrf_exempt
-@staff_member_required
-def remove_tag(request, course_short_title=None):
-    tag = request.POST['tag']
-    object_id = request.POST['object_id']
-    content_type_id = request.POST['content_type_id']
-
-    content_type = ContentType.objects.get_for_id(content_type_id)
-    taggable_object = content_type.get_object_for_this_type(pk=object_id)
-    taggable_object.remove_tag(tag)
-
-    return render_to_response('tags.html', {'tagged_object': taggable_object}, context_instance=RequestContext(request))
-
 
 @csrf_exempt
 @staff_member_required
